Add Web Auth login and logout PoC [SDK-3232] (#14)
* Add Auth0 dependency

* Run `flutter analyze`

* Extract method constants into enums

* Extract the handler's registration

* Add method stubs

* Implement login and logout PoC

+++ auth0_flutter_platform_interface/lib/src/auth/auth_renew_access_token_result.dart
@@ -2,7 +2,7 @@ class AuthRenewAccessTokenResult {
   final String idToken;
   final String accessToken;
   final String? refreshToken;
-  final int expiresIn;
+  final double expiresIn;
   final Set<String> scopes;
 
   AuthRenewAccessTokenResult(
