+++ auth0_flutter_platform_interface/lib/src/credentials.dart
@@ -2,7 +2,7 @@ class Credentials {
   final String idToken;
   final String accessToken;
   final String? refreshToken;
-  final int expiresIn;
+  final double expiresIn;
   final Set<String> scopes;
 
   const Credentials(
