+++ auth0_flutter_platform_interface/lib/src/method_channel_auth0_flutter_auth.dart
@@ -46,7 +46,7 @@ class MethodChannelAuth0FlutterAuth extends Auth0FlutterAuthPlatform {
       idToken: result['idToken'] as String,
       accessToken: result['accessToken'] as String,
       refreshToken: result['refreshToken'] as String,
-      expiresIn: result['expiresIn'] as int,
+      expiresIn: result['expiresIn'] as double,
       scopes: (result['scopes'] as List<Object?>).toTypedSet<String>(),
     );
   }
@@ -82,7 +82,7 @@ class MethodChannelAuth0FlutterAuth extends Auth0FlutterAuthPlatform {
       idToken: result['idToken'] as String,
       accessToken: result['accessToken'] as String,
       refreshToken: result['refreshToken'] as String,
-      expiresIn: result['expiresIn'] as int,
+      expiresIn: result['expiresIn'] as double,
       scopes: (result['scopes'] as List<Object?>).toTypedSet<String>(),
     );
   }
